Combine ChromeBrowser plus input and frame logic into ChromeAudit
We had this strange ChromeBrowser object that was used for finding browser
and renderer threads, but then there were global functions that were used
for finding latency events and frame timing data.

This starts moving these to be methods and sub objects dangling off of the
ChromeAudit object. A followup patch will blow the audit file into pieces.